Moved loss functions from training to layers
