+++ pydrobert/torch/training.py
@@ -12,7 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-'''Functions and classes involved in training'''
+'''Functions and classes involved in training
+
+Notes
+-----
+The loss functions :class:`HardOptimalCompletionDistillationLoss` and
+:class:`MinimumErrorRateLoss` have been moved to :mod:`pydrobert.torch.layers`
+'''
 
 from __future__ import absolute_import
 from __future__ import division
@@ -30,386 +36,15 @@ import torch
 import param
 import pydrobert.torch
 
-from pydrobert.torch.util import error_rate, optimal_completion
 
 __author__ = "Sean Robertson"
 __email__ = "sdrobert@cs.toronto.edu"
 __license__ = "Apache 2.0"
 __copyright__ = "Copyright 2019 Sean Robertson"
 __all__ = [
-    'HardOptimalCompletionDistillationLoss',
-    'MinimumErrorRateLoss',
     'TrainingStateParams',
     'TrainingStateController',
 ]
-
-
-class HardOptimalCompletionDistillationLoss(torch.nn.Module):
-    r'''A categorical loss function over optimal next tokens
-
-    Optimal Completion Distillation (OCD) [sabour2018]_ tries to minimize the
-    train/test discrepancy in transcriptions by allowing seq2seq models to
-    generate whatever sequences they want, then assigns a per-step loss
-    according to whatever next token would set the model on a path that
-    minimizes the edit distance in the future.
-
-    In its "hard" version, the version used in the paper, the OCD loss function
-    is simply a categorical cross-entropy loss of each hypothesis token's
-    distribution versus those optimal next tokens, averaged over the number of
-    optimal next tokens:
-
-    .. math::
-
-        loss(logits_t) = \frac{-\log Pr(s_t|logits_t)}{|S_t|}
-
-    Where :math:`s_t \in S_t` are tokens from the set of optimal next tokens
-    given :math:`hyp_{\leq t}` and `ref`. The loss is decoupled from an exact
-    prefix of `ref`, meaning that `hyp` can be longer or shorter than `ref`.
-
-    When called, this loss function has the signature::
-
-        loss(logits, ref, hyp)
-
-    `hyp` is a long tensor of shape ``(max_hyp_steps, batch_size)`` if
-    `batch_first` is :obj:`False`, otherwise ``(batch_size, max_hyp_steps)``
-    that provides the hypothesis transcriptions. Likewise, `ref` of shape
-    ``(max_ref_steps, batch_size)`` or ``(batch_size, max_ref_steps)``
-    providing reference transcriptions. `logits` is a 4-dimensional tensor of
-    shape ``(max_hyp_steps, batch_size, num_classes)`` if `batch_first` is
-    :obj:`False`, ``(batch_size, max_hyp_steps, num_classes)`` otherwise. A
-    softmax over the step dimension defines the per-step distribution over
-    class labels.
-
-    Parameters
-    ----------
-    eos : int, optional
-        A special token in `ref` and `hyp` whose first occurrence in each
-        batch indicates the end of a transcript
-    include_eos : bool, optional
-        Whether to include the first instance of `eos` found in both `ref` and
-        `hyp` as valid tokens to be computed as part of the distance. `eos`
-        must be a valid class index if `include_eos` is :obj:`True`
-    batch_first : bool, optional
-        Whether the batch dimension comes first, or the step dimension
-    ins_cost : float, optional
-        The cost of an adding a superfluous token to a transcript in `hyp`
-    del_cost : float, optional
-        The cost of missing a token from `ref`
-    sub_cost : float, optional
-        The cost of swapping a token from `ref` with one from `hyp`
-    weight : torch.FloatTensor, optional
-        A manual rescaling weight given to each class
-    reduction : {'mean', 'none', 'sum'}, optional
-        Specifies the reduction to be applied to the output. 'none': no
-        reduction will be applied. 'sum': the output will be summed. 'mean':
-        the output will be averaged.
-
-    Attributes
-    ----------
-    eos : int
-    include_eos, batch_first : bool
-    ins_cost, del_cost, sub_cost : float
-    reduction : {'mean', 'none', 'sum'}
-    weight : torch.FloatTensor or None
-
-    See Also
-    --------
-    pydrobert.torch.util.optimal_completion
-        Used to determine the optimal next token set :math:`S`
-    pydrobert.torch.util.random_walk_advance
-        For producing a random `hyp` based on `logits` if the underlying
-        model producing `logits` is auto-regressive. Also provides an example
-        of sampling non-auto-regressive models
-    '''
-
-    def __init__(
-            self, eos=None, include_eos=True, batch_first=False, ins_cost=1.,
-            del_cost=1., sub_cost=1., weight=None, reduction='mean'):
-        super(HardOptimalCompletionDistillationLoss, self).__init__()
-        self.eos = eos
-        self.include_eos = include_eos
-        self.batch_first = batch_first
-        self.ins_cost = ins_cost
-        self.del_cost = del_cost
-        self.sub_cost = sub_cost
-        self.reduction = reduction
-        self._cross_ent = torch.nn.CrossEntropyLoss(
-            weight=weight, reduction='none'
-        )
-
-    @property
-    def weight(self):
-        return self._cross_ent.weight
-
-    @weight.setter
-    def weight(self, value):
-        self._cross_ent.weight = value
-
-    def check_input(self, logits, ref, hyp):
-        '''Check if input formatted correctly, otherwise RuntimeError'''
-        if logits.dim() != 3:
-            raise RuntimeError('logits must be 3 dimensional')
-        if logits.shape[:-1] != hyp.shape:
-            raise RuntimeError('first two dims of logits must match hyp shape')
-        if self.include_eos and self.eos is not None and (
-                (self.eos < 0) or (self.eos >= logits.shape[-1])):
-            raise RuntimeError(
-                'if include_eos=True, eos ({}) must be a class idx'.format(
-                    self.eos))
-        if self.reduction not in {'mean', 'sum', 'none'}:
-            raise RuntimeError(
-                '"{}" is not a valid value for reduction'
-                ''.format(self.reduction))
-
-    def forward(self, logits, ref, hyp, warn=True):
-        self.check_input(logits, ref, hyp)
-        num_classes = logits.shape[-1]
-        # the padding we use will never be exposed to the user, so we merely
-        # ensure we're not trampling the eos
-        padding = -2 if self.eos == -1 else -1
-        self._cross_ent.ignore_index = padding
-        optimals = optimal_completion(
-            ref, hyp, eos=self.eos, include_eos=self.include_eos,
-            batch_first=self.batch_first, ins_cost=self.ins_cost,
-            del_cost=self.del_cost, sub_cost=self.sub_cost,
-            padding=padding, exclude_last=True, warn=warn,
-        )
-        max_unique_next = optimals.shape[-1]
-        logits = logits.unsqueeze(2).expand(-1, -1, max_unique_next, -1)
-        logits = logits.contiguous()
-        loss = self._cross_ent(
-            logits.view(-1, logits.shape[-1]), optimals.flatten()
-        ).view_as(optimals)
-        padding_mask = optimals.eq(padding)
-        no_padding_mask = padding_mask.eq(0)
-        loss = loss.masked_fill(padding_mask, 0.).sum(2)
-        loss = torch.where(
-            no_padding_mask.any(2),
-            loss / no_padding_mask.float().sum(2),
-            loss,
-        )
-        if self.reduction == 'mean':
-            loss = loss.mean()
-        elif self.reduction == 'sum':
-            loss = loss.sum()
-        return loss
-
-
-class MinimumErrorRateLoss(torch.nn.Module):
-    r'''Error rate expectation normalized over some number of transcripts
-
-    Proposed in [prabhavalkar2018]_ though similar ideas had been explored
-    previously. Given a subset of all possible token sequences and their
-    associated probability mass over that population, this loss calculates the
-    probability mass normalized over the subset, then calculates the
-    expected error rate over that normalized distribution. That is, given some
-    sequences :math:`s \in S \subseteq P`, the loss for a given reference
-    transcription :math:`s^*` is
-
-    .. math::
-
-        \mathcal{L}(s, s^*) = \frac{Pr(s) ER(s, s^*)}{\sum_{s'} Pr(s')}
-
-    This is an exact expectation over :math:`S` but not over :math:`P`. The
-    larger the mass covered by :math:`S`, the closer the expectation is to the
-    population - especially so for an n-best list (though it would be biased).
-
-    This loss function has the following signature::
-
-        loss(log_probs, ref, hyp)
-
-    `log_probs` is a tensor of shape ``(batch_size, samples)`` providing the
-    log joint probabilities of every path. `hyp` is a long tensor of shape
-    ``(max_hyp_steps, batch_size, samples)`` if `batch_first` is :obj:`False`
-    otherwise ``(batch_size, samples, max_hyp_steps)`` that provides the
-    hypothesis transcriptions. `ref` is a 2- or 3-dimensional tensor. If 2D, it
-    is of shape ``(max_ref_steps, batch_size)`` (or ``(batch_size,
-    max_ref_steps)``). Alternatively, `ref` can be of shape ``(max_ref_steps,
-    batch_size, samples)`` or ``(batch_size, samples, max_ref_steps)``.
-
-    If `ref` is 2D, the loss is calculated as
-
-    .. math::
-
-        loss_{MER} = SoftMax(log\_probs)[ER(hyp_i, ref) - \mu_i]
-
-    where :math:`\mu_i` is the average error rate along paths in the batch
-    element :math:`i`. :math:`mu_i` can be removed by setting `sub_avg` to
-    :obj:`False`. Note that each hypothesis is compared against the same
-    reference as long as the batch element remains the same
-
-    If `ref` is 3D, the loss is calculated as
-
-    .. math::
-
-        loss_{MER} = SoftMax(log\_probs)[ER(hyp_i, ref_i) - \mu_i]
-
-    In this version, each hypothesis is compared against a unique reference
-
-    Parameters
-    ----------
-    eos : int, optional
-        A special token in `ref` and `hyp` whose first occurrence in each
-        batch indicates the end of a transcript
-    include_eos : bool, optional
-        Whether to include the first instance of `eos` found in both `ref` and
-        `hyp` as valid tokens to be computed as part of the distance.
-    sub_avg : bool, optional
-        Whether to subtract the average error rate from each pathwise error
-        rate
-    batch_first : bool, optional
-        Whether batch/path dimensions come first, or the step dimension
-    norm : bool, optional
-        If :obj:`False`, will use edit distances instead of error rates
-    ins_cost : float, optional
-        The cost of an adding a superfluous token to a transcript in `hyp`
-    del_cost : float, optional
-        The cost of missing a token from `ref`
-    sub_cost : float, optional
-        The cost of swapping a token from `ref` with one from `hyp`
-    reduction : {'mean', 'none', 'sum'}, optional
-        Specifies the reduction to be applied to the output. 'none': no
-        reduction will be applied. 'sum': the output will be summed. 'mean':
-        the output will be averaged.
-
-    Attributes
-    ----------
-    eos, ignore_index : int
-    include_eos, sub_avg, batch_first, norm : bool
-    ins_cost, del_cost, sub_cost : float
-    reduction : {'mean', 'none', 'sum'}
-
-    Notes
-    -----
-
-    A previous version of this module incorporated a Maximum Likelihood
-    Estimate (MLE) into the loss as in [prabhavalkar2018]_, which required
-    `logits` instead of `log_probs`. This was overly complicated, given the
-    user can easily incorporate the additional loss term herself by using
-    :class:`torch.nn.CrossEntropyLoss`. Take a look at the example below for
-    how to recreate this
-
-    Examples
-    --------
-
-    Assume here that `logits` is the output of some neural network, and that
-    `hyp` has somehow been produced from that (e.g. a beam search or random
-    walk). We combine this loss function with a cross-entropy/MLE term to
-    sort-of recreate [prabhavalkar2018]_.
-
-    >>> from pydrobert.torch.util import sequence_log_probs
-    >>> steps, batch_size, num_classes, eos, padding = 30, 20, 10, 0, -1
-    >>> samples, lmb = 10, .01
-    >>> logits = torch.randn(
-    ...     steps, samples, batch_size, num_classes, requires_grad=True)
-    >>> hyp = torch.randint(num_classes, (steps, samples, batch_size))
-    >>> ref_lens = torch.randint(1, steps + 1, (batch_size,))
-    >>> ref_lens[0] = steps
-    >>> ref = torch.nn.utils.rnn.pad_sequence(
-    ...     [torch.randint(1, num_classes, (x,)) for x in ref_lens],
-    ...     padding_value=padding,
-    ... )
-    >>> ref[ref_lens - 1, range(batch_size)] = eos
-    >>> ref = ref.unsqueeze(1).repeat(1, samples, 1)
-    >>> mer = MinimumErrorRateLoss(eos=eos)
-    >>> mle = torch.nn.CrossEntropyLoss(ignore_index=padding)
-    >>> log_probs = sequence_log_probs(logits, hyp, eos=eos)
-    >>> l = mer(log_probs, ref, hyp)
-    >>> l = l + lmb * mle(logits.view(-1, num_classes), ref.flatten())
-    >>> l.backward()
-
-    See Also
-    --------
-    pydrobert.torch.util.beam_search_advance
-        For getting an n-best list into `hyp` and some `log_probs`.
-    pydrobert.torch.util.random_walk_advance
-        For getting a random sample into `hyp`
-    pydrobert.torch.util.sequence_log_probs
-        For converting token log probs (or logits) to sequence log probs
-    '''
-
-    def __init__(
-            self, eos=None, include_eos=True, sub_avg=True, batch_first=False,
-            norm=True, ins_cost=1., del_cost=1., sub_cost=1.,
-            reduction='mean'):
-        super(MinimumErrorRateLoss, self).__init__()
-        self.eos = eos
-        self.include_eos = include_eos
-        self.sub_avg = sub_avg
-        self.batch_first = batch_first
-        self.norm = norm
-        self.ins_cost = ins_cost
-        self.del_cost = del_cost
-        self.sub_cost = sub_cost
-        self.reduction = reduction
-
-    def check_input(self, log_probs, ref, hyp):
-        '''Check if the input is formatted correctly, otherwise RuntimeError'''
-        if log_probs.dim() != 2:
-            raise RuntimeError('log_probs must be 2 dimensional')
-        if hyp.dim() != 3:
-            raise RuntimeError('hyp must be 3 dimensional')
-        if ref.dim() not in {2, 3}:
-            raise RuntimeError('ref must be 2 or 3 dimensional')
-        if self.batch_first:
-            if ref.dim() == 2:
-                ref = ref.unsqueeze(1).expand(-1, hyp.shape[1], -1)
-            if (
-                    (ref.shape[:2] != hyp.shape[:2]) or
-                    (ref.shape[:2] != log_probs.shape)):
-                raise RuntimeError(
-                    'ref and hyp batch_size and sample dimensions must match')
-            if ref.shape[1] < 2:
-                raise RuntimeError(
-                    'Batch must have at least two samples, got {}'
-                    ''.format(ref.shape[1]))
-        else:
-            if ref.dim() == 2:
-                ref = ref.unsqueeze(-1).expand(-1, -1, hyp.shape[-1])
-            if (
-                    (ref.shape[1:] != hyp.shape[1:]) or
-                    (ref.shape[1:] != log_probs.shape)):
-                raise RuntimeError(
-                    'ref and hyp batch_size and sample dimensions must match')
-            if ref.shape[2] < 2:
-                raise RuntimeError(
-                    'Batch must have at least two samples, got {}'
-                    ''.format(ref.shape[2]))
-        if self.reduction not in {'mean', 'sum', 'none'}:
-            raise RuntimeError(
-                '"{}" is not a valid value for reduction'
-                ''.format(self.reduction))
-
-    def forward(self, log_probs, ref, hyp, warn=True):
-        self.check_input(log_probs, ref, hyp)
-        if self.batch_first:
-            batch_size, samples, max_hyp_steps = hyp.shape
-            max_ref_steps = ref.shape[-1]
-            if ref.dim() == 2:
-                ref = ref.unsqueeze(1).repeat(1, samples, 1)
-            ref = ref.view(-1, max_ref_steps)
-            hyp = hyp.view(-1, max_hyp_steps)
-        else:
-            max_hyp_steps, batch_size, samples = hyp.shape
-            max_ref_steps = ref.shape[0]
-            if ref.dim() == 2:
-                ref = ref.unsqueeze(-1).repeat(1, 1, samples)
-            ref = ref.view(max_ref_steps, -1)
-            hyp = hyp.view(max_hyp_steps, -1)
-        er = error_rate(
-            ref, hyp, eos=self.eos, include_eos=self.include_eos,
-            norm=self.norm, batch_first=self.batch_first,
-            ins_cost=self.ins_cost, del_cost=self.del_cost,
-            sub_cost=self.sub_cost, warn=warn,
-        ).view(batch_size, samples)
-        if self.sub_avg:
-            er = er - er.mean(1, keepdim=True)
-        loss = er * torch.nn.functional.softmax(log_probs, 1)
-        if self.reduction == 'mean':
-            loss = loss.mean()
-        elif self.reduction == 'sum':
-            loss = loss.sum()
-        return loss
 
 
 class TrainingStateParams(param.Parameterized):
