+++ tests/test_layers.py
@@ -12,6 +12,133 @@ __license__ = "Apache 2.0"
 __copyright__ = "Copyright 2019 Sean Robertson"
 
 
+@pytest.mark.parametrize('batch_first', [True, False])
+@pytest.mark.parametrize('sub_avg', [True, False])
+@pytest.mark.parametrize('reduction', ['mean', 'none'])
+def test_minimum_error_rate_loss(device, batch_first, sub_avg, reduction):
+    torch.manual_seed(100)
+    num_batches, samples, num_classes = 5, 5, 30
+    max_ref_steps, max_hyp_steps = 10, 5
+    assert max_ref_steps > max_hyp_steps  # nonzero loss guaranteed
+    if batch_first:
+        hyp = torch.randint(
+            num_classes, (num_batches, samples, max_hyp_steps), device=device)
+        hyp[..., 0] = 0
+        ref = torch.randint(
+            num_classes, (num_batches, max_ref_steps), device=device)
+        ref[..., 0] = 0
+    else:
+        hyp = torch.randint(
+            num_classes, (max_hyp_steps, num_batches, samples), device=device)
+        hyp[0] = 0
+        ref = torch.randint(
+            num_classes, (max_ref_steps, num_batches), device=device)
+        ref[0] = 0
+    log_probs = torch.randn(num_batches, samples, device=device)
+    loss = layers.MinimumErrorRateLoss(
+        eos=None, sub_avg=sub_avg, batch_first=batch_first,
+        reduction=reduction,
+    )
+    l1 = loss(log_probs, ref, hyp)
+    assert l1.ne(0.).any()
+    l2 = loss(log_probs, ref, hyp)
+    assert torch.allclose(l1, l2)
+    loss.eos = 0
+    l3 = loss(log_probs, ref, hyp)
+    assert l3.eq(0.).all()
+
+
+@pytest.mark.parametrize('batch_first', [True, False])
+@pytest.mark.parametrize('eos', [None, 0])
+@pytest.mark.parametrize('ref_steps_times', [1, 2])
+@pytest.mark.parametrize('reduction', ['mean', 'none'])
+@pytest.mark.parametrize('include_eos', [True, False])
+def test_hard_optimal_completion_distillation_loss(
+        device, batch_first, eos, ref_steps_times, reduction, include_eos):
+    torch.manual_seed(209384)
+    num_batches, max_steps, num_classes = 20, 41, 10
+    if eos is None:
+        hyp_lens = torch.tensor(max_steps).expand(num_batches)
+        ref_lens = torch.tensor(ref_steps_times * max_steps)
+        ref_lens = ref_lens.expand(num_batches)
+    else:
+        hyp_lens = torch.randint(1, max_steps + 1, (num_batches,))
+        ref_lens = torch.randint(
+            2, ref_steps_times * max_steps + 1, (num_batches,))
+    ref = torch.nn.utils.rnn.pad_sequence(
+        [
+            torch.randint(1, num_classes, (x,))
+            for x in ref_lens
+        ],
+        padding_value=num_classes - 1, batch_first=batch_first,
+    )
+    hyp = torch.nn.utils.rnn.pad_sequence(
+        [
+            torch.randint(1, num_classes, (x,))
+            for x in hyp_lens
+        ],
+        padding_value=-1, batch_first=batch_first,
+    )
+    if eos is not None:
+        for bt in range(num_batches):
+            if batch_first:
+                ref[bt, ref_lens[bt] - 1] = eos
+                hyp[bt, hyp_lens[bt] - 1] = eos
+            else:
+                ref[ref_lens[bt] - 1, bt] = eos
+                hyp[hyp_lens[bt] - 1, bt] = eos
+        if not include_eos:
+            ref_lens = ref_lens - 1
+            hyp_lens = hyp_lens - 1
+    logits = torch.rand(tuple(hyp.shape) + (num_classes,))
+    if batch_first:
+        len_mask = (
+            torch.arange(hyp.shape[1]).unsqueeze(0) < hyp_lens.unsqueeze(1)
+        )
+    else:
+        len_mask = torch.arange(hyp.shape[0]).unsqueeze(1) < hyp_lens
+    logits, ref, hyp = logits.to(device), ref.to(device), hyp.to(device)
+    ref_lens, hyp_lens = ref_lens.to(device), hyp_lens.to(device)
+    len_mask = len_mask.to(device)
+    inv_len_mask = len_mask.eq(0)
+    logits.requires_grad_(True)
+    loss = layers.HardOptimalCompletionDistillationLoss(
+        eos=eos, include_eos=include_eos, batch_first=batch_first,
+        reduction=reduction,
+    )
+    l1 = loss(logits, ref, hyp)
+    assert torch.all(l1 == l1)  # no nans
+    if reduction == 'none':
+        assert torch.all(l1.masked_select(inv_len_mask).eq(0.))
+        # reference transcriptions are all positive length, so the first
+        # optimal completion (assuming hyp length is nonzero) will always be
+        # the first token in ref (and only the first token), given that there's
+        # no ambiguity in the alignment of the prefix ""
+        log_probs = torch.nn.functional.log_softmax(logits, 2)
+        if batch_first:
+            zero_length_mask = ref_lens.eq(0).unsqueeze(1)
+            first_loss = torch.where(
+                zero_length_mask,
+                torch.zeros_like(log_probs[:, 0, 0]),
+                -log_probs[:, 0].gather(
+                    1, ref[:, 0].unsqueeze(-1)
+                ).squeeze(-1),
+            )
+            assert torch.allclose(l1[:, 0], first_loss)
+        else:
+            zero_length_mask = ref_lens.eq(0).unsqueeze(0)
+            first_loss = torch.where(
+                zero_length_mask,
+                torch.zeros_like(log_probs[0, :, 0]),
+                -log_probs[0].gather(1, ref[0].unsqueeze(-1)).squeeze(-1),
+            )
+            assert torch.allclose(l1[0], first_loss)
+        l1 = l1.mean()
+    g, = torch.autograd.grad([l1], [logits])
+    assert torch.all(g.masked_select(inv_len_mask.unsqueeze(-1)).eq(0.))
+    assert not torch.all(g.eq(0.))
+
+
 @pytest.mark.parametrize('dim', [0, 1])
 def test_global_soft_attention(device, dim):
